requirements.txt
flask~=3.0.3
Flask-SQLAlchemy
SQLAlchemy~=2.0.34
Pymysql
python-dotenv
cryptography